UX Refactoring
Moved out UX rendering such as CSS and logo into utility functions
