+++ app.py
@@ -3,7 +3,6 @@ import sys
 import sqlite3
 import importlib.util
 import streamlit as st
-import streamlit.components.v1 as components
 
 import html
 import logging
@@ -16,6 +15,7 @@ from db_queries.industries import get_industries
 from db_queries.growth_stages import determine_company_stage
 from db_queries.connection import get_db_connection
 from utils.slider_helpers import get_slider_format, get_slider_range, get_step_size
+from utils.ux_helpers import add_toolbar, add_logo, load_css, load_js
 
 
 # Load and use Streamlit config
@@ -127,84 +127,6 @@ except Exception as e:
     quaternary_color = "#4D466B"  # Purple-gray
     highlight_color = "#AC2147"  # Red highlight
     link_color = "#00A8A8"  # Teal for hyperlinks
-
-
-# Load CSS Dynamically
-def load_css(file_path):
-    with open(file_path) as f:
-        st.markdown(f"<style>{f.read()}</style>", unsafe_allow_html=True)
-
-
-# Load Javascript dynamically
-def load_js(file_path):
-    with open(file_path) as f:
-        js_code = f.read()
-        components.html(f"<script>{js_code}</script>", height=0)
-
-
-def add_toolbar():
-    st.markdown("""
-    <style>
-        /* Hide Streamlit's default header and menu */
-        header, #MainMenu, footer {visibility: hidden;}
-    
-        /* Remove top margin/padding */
-        .block-container { padding-top: 0 !important; margin-top: 0 !important; }
-        .css-18e3th9 { padding-top: 0 !important; } /* Sometimes needed for newer Streamlit */
-    </style>
-    """, unsafe_allow_html=True)
-
-    app_toolbar_html = f"""
-    <div class="top-toolbar">
-        <span class="toolbar-text">Need Clarity?</span>
-        <a href="https://outlook.office.com/owa/calendar/MinimalistInnovationLLC@minimalistinnovation.onmicrosoft.com/bookings/s/H_o18Z1ej0OAvMiMMMyhTA2" 
-            class="toolbar-cta" 
-            target="_blank">
-            🤙Schedule a Call Now
-        </a>
-    </div>
-    <div class="toolbar-spacer"></div>
-    """
-    st.markdown(app_toolbar_html, unsafe_allow_html=True)
-
-
-# === Helper Functions ===
-def add_logo():
-    try:
-        # Create a container with two columns
-        col1, col2 = st.columns([3, 1])
-
-        with col1:
-            # Left side: App logo (existing)
-            app_logo_html = f"""
-            <div style="display: flex; align-items: center;">
-                <div style="background-color: {primary_color}; width: 40px; height: 40px; border-radius: 8px; 
-                          display: flex; justify-content: center; align-items: center; margin-right: 12px;">
-                    <span style="color: white; font-weight: bold; font-size: 18px;">ATA</span>
-                </div>
-                <div>
-                    <h2 style="margin: 0; padding: 0; color: {primary_color};">Adaptive Traction Architecture</h2>
-                    <h3 style="margin: 0; padding: 0; color: {primary_color};"> Diagnostics</h3>
-                </div>
-            </div>
-            """
-            st.markdown(app_logo_html, unsafe_allow_html=True)
-
-        with col2:
-            # Right side: Company logo using st.image
-            import os
-            from PIL import Image
-
-            # Path to your SVG logo
-            logo_path = os.path.join("media", "Minimalist_Horizontal_Blue.svg")
-
-            # Display the logo - align it to the right
-            st.image(logo_path, width=200)  # Adjust width as needed
-
-        logger.debug("Logos rendered successfully")
-    except Exception as e:
-        logger.error(f"Error displaying logos: {str(e)}")
-        st.error(f"Error displaying logos: {str(e)}")
 
 
 def import_and_setup_database():
@@ -337,7 +259,7 @@ def main():
         # Add the toolbar
         add_toolbar()
         # Add logo
-        add_logo()
+        add_logo(primary_color)
 
         # Basic interactivity for demo purposes
         st.markdown("<div class='dashboard-container'>", unsafe_allow_html=True)
